refactor: extract roamer tracking logic into reusable utility
