+++ src/engine/assistant/strategies/gen2Strategy.ts
@@ -3,6 +3,7 @@ import { getGen2UnobtainableReason } from '../../exclusives/gen2Exclusives';
 import { getDistanceToMap, resolveOutdoorMapId } from '../../mapGraph/gen2Graph';
 import type { SaveData } from '../../saveParser/index';
 import type { AssistantStrategy, Suggestion } from './types';
+import { getRoamerSuggestions } from './utils/roamer';
 
 export const gen2Strategy: AssistantStrategy = {
   generation: 2,
@@ -24,36 +25,12 @@ export const gen2Strategy: AssistantStrategy = {
     const missingSet = new Set(missingIds);
 
     // 1. Roamer tracking
-    // ⚡ Bolt: converted roamingLegendaries to a Map keyed by speciesId to optimize O(N) loop lookup
-    const roamerMap = new Map(saveData.roamingLegendaries?.map((r) => [r.speciesId, r]));
     const roamers = [
       { id: 243, name: 'Raikou' },
       { id: 244, name: 'Entei' },
       { id: 245, name: 'Suicune' },
     ];
-    for (const roamer of roamers) {
-      // In Crystal, Suicune is a static encounter, not a roamer. Suppress roamer logic for it.
-      if (roamer.id === 245 && saveData.gameVersion === 'crystal') continue;
-
-      if (missingSet.has(roamer.id)) {
-        let isTracked = false;
-        const rData = roamerMap.get(roamer.id);
-        if (rData && rData.mapId !== 0) {
-          isTracked = true;
-        }
-
-        suggestions.push({
-          id: `roamer-${roamer.id}`,
-          category: 'Catch',
-          title: `Track ${roamer.name}`,
-          description: isTracked
-            ? `${roamer.name} is currently roaming! Check your Pokédex to see its current route.`
-            : `Encounter ${roamer.name} in the wild, then use your Pokédex to track its location!`,
-          pokemonId: roamer.id,
-          priority: 85,
-        });
-      }
-    }
+    suggestions.push(...getRoamerSuggestions(saveData, missingSet, roamers, saveData.gameVersion === 'crystal'));
 
     // 2. Headbutt / Rock Smash
     // TM08 (Rock Smash) is Item ID 198 (0xC6), TM02 (Headbutt) is Item ID 192 (0xC0)
